Fix compile errors on Win64. Review URL: https://chromiumcodereview.appspot.com/10913273
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@12509 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/heap.cc
@@ -680,7 +680,7 @@ static bool AbortIncrementalMarkingAndCollectGarbage(
 
 
 void Heap::ReserveSpace(
-    intptr_t *sizes,
+    int *sizes,
     Address *locations_out) {
   bool gc_performed = true;
   int counter = 0;
